chore: add tests for listeners
