+++ tests/mocha/index.html
@@ -106,6 +106,8 @@
       import './metrics_test.js';
       import './mutator_test.js';
       import './names_test.js';
+      // TODO: Remove these tests.
+      import './old_workspace_comment_test.js';
       import './procedure_map_test.js';
       import './blocks/procedures_test.js';
       import './registry_test.js';
